Graph elements refactor moved picker and total spent label to its own swift file
